✨  feat : ValueService와 ValueQueryService 분리
