+++ src/main/java/org/likelion/zagabi/domain/Value/Dto/request/UpdateValueRequestDto.java
@@ -4,7 +4,6 @@ import lombok.Builder;
 import lombok.Getter;
 
 @Builder
-@Getter
 public record UpdateValueRequestDto(
         Long valueId,
         Integer changeRanking
