+++ src/main/resources/application.yml
@@ -9,7 +9,7 @@ spring:
   jpa:
      open-in-view: true
      hibernate:
-      ddl-auto: update
+      ddl-auto: create
      show-sql: true
